feat(desk) move some app toolbar logic to custom hook
